Fix building functionality.
A lack of quotation marks in internals/thirdparty_helper/build_thirdparty.sh led to invalid logic
in locating the base directory.

Instead of the if statement registering False if the user was not runnning building in a slurm job, it would evaluate to True and attempt to proceed as if the user had sbatch'd the build logic.

This would lead to a downstream error preventing building due to inability to locate base directory.

Addition of quotation marks leads to correct location of base directory.